add price new

+++ src/main/java/com/example/udpm14sellcomputerpartsbackend/service/impl/CartServiceImpl.java
@@ -112,7 +112,7 @@ public class CartServiceImpl implements CartService {
                 cart.setProductId(productEntity.getId());
                 cart.setImage(imageEntity.get(0).getLink());
                 cart.setPrice(productEntity.getDiscount());
-                cart.setTotal(productEntity.getDiscount() * 1);
+                cart.setTotal((long) (productEntity.getDiscount() * 1));
                 cart.setQuantity(1);
             } else {//Neu san pham da co trong database tang so luong them 1
                 cart.setQuantity(cart.getQuantity() + 1);
